pyproject.toml
[tool.poetry]
name = "t-rag"
version = "0.1.0"
description = ""
authors = ["matteonotaro <matteo.notaro@agilelab.it>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
sqlalchemy-bigquery = "^1.12.0"
llama-index = "^0.12.8"
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
